setting: storybook 설치 및 적용
