+++ package.json
@@ -10,7 +10,9 @@
     "lint:staged": "lint-staged",
     "lint:css": "stylelint '**/*.{tsx, ts}' --fix",
     "prepare": "husky install",
-    "vercel-deploy": "next build && next export"
+    "vercel-deploy": "next build && next export",
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build"
   },
   "husky": {
     "hooks": {
@@ -59,6 +61,14 @@
     "typescript": "5.1.3"
   },
   "devDependencies": {
+    "@storybook/addon-essentials": "^7.4.2",
+    "@storybook/addon-interactions": "^7.4.2",
+    "@storybook/addon-links": "^7.4.2",
+    "@storybook/addon-onboarding": "^1.0.8",
+    "@storybook/blocks": "^7.4.2",
+    "@storybook/nextjs": "^7.4.2",
+    "@storybook/react": "^7.4.2",
+    "@storybook/testing-library": "^0.2.1",
     "@types/js-cookie": "^3.0.3",
     "@types/react-slick": "^0.23.10",
     "@types/styled-components": "^5.1.26",
@@ -71,7 +81,9 @@
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-storybook": "^0.6.13",
     "prettier": "^2.8.8",
+    "storybook": "^7.4.2",
     "stylelint-config-concentric-order": "^5.1.0",
     "stylelint-config-standard": "^33.0.0",
     "stylelint-processor-styled-components": "^1.10.0"
